feature(AWS Lambda): added java sample fucntion for sync direct
