+++ aws_lambda/.gitignore
@@ -1,2 +1,15 @@
+## go executables
+**/main
+
+## java executabels
+**/target/
+
+# vscode files
+**/.DS_Store
+.DS_Store
+
+# intellij files
+**/.idea
+
 *.zip
-/main
+**/*.zip
